Deduplicate feed metadata and generation code
